Module ncollide::bounding_volume

Bounding volumes.

Structs

AABB
An Axis Aligned Bounding Box.

BoundingSphere
A Bounding Sphere.

BoundingVolumeInterferencesCollector
Bounding Volume Tree visitor collecting interferences with a given bounding volume.

Traits

BoundingVolume
Trait of bounding volumes.

HasBoundingVolume
Traits of objects having a bounding volume.

Functions

aabb
Computes the axis-aligned bounding box of a shape g transformed by m.

ball_aabb
Computes the Axis-Aligned Bounding Box of a ball.

bounding_sphere
Computes the bounding sphere of a shape g transformed by m.

implicit_shape_aabb
Computes the AABB of an support mapped shape.

point_cloud_aabb
Computes the AABB of a set of point.

point_cloud_bounding_sphere
Computes a bounding sphere of the specified set of point.

point_cloud_bounding_sphere_with_center
Computes the bounding sphere of a set of point, given its center.

Type Definitions

AABB2

AABB3

BoundingSphere2

BoundingSphere3