Module ncollide::broad_phase

Broad phases.

Structs

BroadPhasePairFilters
Filters deciding whether a proximity is to be further investigated by the narrow phase or not.

BruteForceBroadPhase
A broad phase testing explicitly all bounding volume pairs.

DBVTBroadPhase
Broad phase based on a Dynamic Bounding Volume Tree.

Traits

BroadPhase
Trait all broad phase must implement.

BroadPhasePairFilter
A signal handler for contact detection.