Module ncollide::geometry::algorithms::gjk

The Gilbert–Johnson–Keerthi distance algorithm.

Enums

GJKResult
Results of the GJK algorithm.

Functions

cast_ray
Casts a ray on a support map using the GJK algorithm.

closest_points
Computes the closest points between two convex shapes unsing the GJK algorithm.

closest_points_with_max_dist
Computes the closest points between two convex shapes unsing the GJK algorithm.

distance
Computes the exact distance separating two convex shapes unsing the GJK. algorithm.

project_origin
Projects the origin on a shape unsing the GJK algorithm.

project_origin_with_max_dist
Projects the origin on a shape using the Separating Axis GJK algorithm. The algorithm will stop as soon as the polytope can be proven to be at least max_dist away from the origin.