Enum ncollide::geometry::algorithms::gjk::GJKResult

pub enum GJKResult<P, V> {
    Intersection,
    Projection(P),
    NoIntersection(V),
}

Results of the GJK algorithm.

Variants

Intersection

Result of the GJK algorithm when the origin is inside of the polytope.

Projection

Result of the GJK algorithm when a projection of the origin on the polytope is found.

NoIntersection

Result of the GJK algorithm when the origin is to far away from the polytope.