Function ncollide::geometry::algorithms::gjk::closest_points

pub fn closest_points<P, M, S, G1, G2>(m1: &M, g1: &G1, m2: &M, g2: &G2, simplex: &mut S) -> Option<(P, P)> where G2: SupportMap<P, M> + ?Sized, P: Point, S: Simplex<AnnotatedPoint<P>>, G1: SupportMap<P, M> + ?Sized

Computes the closest points between two convex shapes unsing the GJK algorithm.

Arguments:

g1 - first shape.

g2 - second shape.

simplex - the simplex to be used by the GJK algorithm. It must be already initialized with at least one point on the shapes CSO. See minkowski_sum::cso_support_point to compute such point.