Function ncollide::geometry::algorithms::gjk::distance

pub fn distance<P, M, S, G1, G2>(m1: &M, g1: &G1, m2: &M, g2: &G2, simplex: &mut S) -> P::Vect::Scalar where S: Simplex<P>, P: Point, G1: SupportMap<P, M> + ?Sized, G2: SupportMap<P, M> + ?Sized

Computes the exact distance separating two convex shapes unsing the GJK. algorithm.

Arguments:

g1 - first shape.

g2 - second shape.

simplex - the simplex to be used by the GJK algorithm. It must be already initialized with at least one point on the shapes CSO. See minkowski_sum::cso_support_point to compute such point.