Function ncollide::geometry::algorithms::gjk::project_origin

pub fn project_origin<P, M, S, G>(m: &M, shape: &G, simplex: &mut S) -> Option<P> where G: SupportMap<P, M> + ?Sized, S: Simplex<P>, P: Point

Projects the origin on a shape unsing the GJK algorithm.

Arguments:

shape - the shape to project the origin on

simplex - the simplex to be used by the GJK algorithm. It must be already initialized with at least one point on the shape boundary.