Function ncollide::geometry::algorithms::minkowski_sampling::closest_points

pub fn closest_points<N, P, V, M, S, G1, G2>(m1: &M, g1: &G1, m2: &M, g2: &G2, simplex: &mut S) -> Option<(P, P, V)> where N: Scalar, P: Point<N, V>, V: Vect<N>, V: Translate<P>, M: Translation<V>, S: Simplex<N, AnnotatedPoint<P>>, G1: SupportMap<P, V, M>, G1: PreferedSamplingDirections<V, M>, G2: SupportMap<P, V, M>, G2: PreferedSamplingDirections<V, M>, G1: ?Sized, G2: ?Sized

Computes the closest points between two implicit inter-penetrating shapes. Returns None if the shapes are not in penetration. This can be used as a fallback algorithm for the GJK algorithm.