Module ncollide::inspection

Traits and methods to inspect and retrieve the capabilities and representations of shapes at runtime.

Structs

ReprDesc

Traits

Repr
An object with a unique runtime geometric representation.

Functions

composite_shape_repr_id
Gets the id associated with the CompositeShape trait.

maybe_as_composite_shape
Converts a shape to a composite shape if possible.

maybe_as_support_map
Converts a shape to a support map if possible.

maybe_repr_desc_as_composite_shape
Converts a shape to a composite shape if possible.

maybe_repr_desc_as_support_map
Converts a shape descriptor to a support map if possible.

support_map_repr_id
Gets the id associated with the SupportMap trait.

Type Definitions

Repr2

Repr3