Struct ncollide::inspection::ReprDesc

pub struct ReprDesc<'a> {
    // some fields omitted
}

Methods

impl<'a> ReprDesc<'a>

unsafe fn new(type_id: TypeId, repr_id: TypeId, repr: TraitObject) -> ReprDesc<'a>

Creates a new representation descriptor.

This is unsafe as there is no way to check that the given triple of data are valid.

fn type_id(&self) -> TypeId

TypeId of this shape's exact type.

fn repr_id(&self) -> TypeId

TypeId of this shape's representation.

fn repr(&self) -> TraitObject

This shape's representation.

fn downcast_ref<T>(&self) -> Option<&T> where T: 'static

Converts this repr as an exact shape.