Struct ncollide::narrow_phase::IncrementalContactManifoldGenerator

pub struct IncrementalContactManifoldGenerator<P, M, CD> where P: Point { /* fields omitted */ }

Contact manifold generator which keeps track of several contacts.

One contact is added per update until the maximum number of contact is reached. When the maximum number of contact is reached, each time a new contact is created, the new manifold is computed by maximizing the variance along each canonical axis (of the space in which leaves the contacts).

Methods

impl<P, M, CD> IncrementalContactManifoldGenerator<P, M, CD> where CD: ContactGenerator<P, M>,
P: Point

fn new(cd: CD) -> IncrementalContactManifoldGenerator<P, M, CD>

Creates a new incremental contact manifold generator.

Arguments:

cd - collision detection sub-algorithm used to generate the contact points.

impl<P, M, CD> IncrementalContactManifoldGenerator<P, M, CD> where CD: ContactGenerator<P, M>,
M: Isometry<P>,
P: Point

fn get_sub_collision(&mut self,
d: &ContactDispatcher<P, M>,
m1: &M,
g1: &Shape<P, M> + 'static,
m2: &M,
g2: &Shape<P, M> + 'static,
prediction: P::Real)
-> Option<Option<Contact<P>>>

Gets a collision from the sub-detector used by this manifold generator. This does not update the manifold itself.

fn add_new_contacts(&mut self,
d: &ContactDispatcher<P, M>,
m1: &M,
g1: &Shape<P, M> + 'static,
m2: &M,
g2: &Shape<P, M> + 'static,
prediction: P::Real)
-> bool

Updates the current manifold by adding one point.

fn update_contacts(&mut self, m1: &M, m2: &M, prediction: P::Real)

Updates the contacts already existing on this manifold.