Struct ncollide::narrow_phase::SupportMapPlaneContactGenerator

pub struct SupportMapPlaneContactGenerator<P, M> where P: Point { /* fields omitted */ }

Collision detector between a plane and a shape implementing the SupportMap trait.

This detector generates only one contact point. For a full manifold generation, see IncrementalContactManifoldGenerator.

Methods

impl<P, M> SupportMapPlaneContactGenerator<P, M> where P: Point

fn new() -> SupportMapPlaneContactGenerator<P, M>

Creates a new persistent collision detector between a plane and a shape with a support mapping function.