Struct ncollide::narrow_phase::SupportMapSupportMapContactGenerator

pub struct SupportMapSupportMapContactGenerator<P, M, S> where P: Point { /* fields omitted */ }

Persistent collision detector between two shapes having a support mapping function.

It is based on the GJK algorithm. This detector generates only one contact point. For a full manifold generation, see IncrementalContactManifoldGenerator.

Methods

impl<P, M, S> SupportMapSupportMapContactGenerator<P, M, S> where P: Point,
S: Simplex<AnnotatedPoint<P>>

fn new(simplex: S) -> SupportMapSupportMapContactGenerator<P, M, S>

Creates a new persistant collision detector between two shapes with support mapping functions.

It is initialized with a pre-created simplex.