Trait ncollide::narrow_phase::CollisionDetector

pub trait CollisionDetector<P, M> where P: Point {
    fn update(&mut self, dispatcher: &CollisionDispatcher<P, M>, ma: &M, a: &Repr<P, M> + 'static, mb: &M, b: &Repr<P, M> + 'static, prediction: P::Vect::Scalar) -> bool;
    fn num_colls(&self) -> usize;
    fn colls(&self, &mut Vec<Contact<P>>);
}

Trait implemented algorithms that compute contact points, normals and penetration depths.

Required Methods

fn update(&mut self, dispatcher: &CollisionDispatcher<P, M>, ma: &M, a: &Repr<P, M> + 'static, mb: &M, b: &Repr<P, M> + 'static, prediction: P::Vect::Scalar) -> bool

Runs the collision detection on two objects. It is assumed that the same collision detector (the same structure) is always used with the same pair of object.

fn num_colls(&self) -> usize

The number of collision detected during the last update.

fn colls(&self, &mut Vec<Contact<P>>)

Collects the collisions detected during the last update.