Trait ncollide::narrow_phase::ContactHandler

pub trait ContactHandler<P, M, T> where P: Point {
    fn handle_contact_started(&mut self,
                              co1: &CollisionObject<P, M, T>,
                              co2: &CollisionObject<P, M, T>,
                              contacts: &Box<ContactGenerator<P, M> + 'static>);
    fn handle_contact_stopped(&mut self,
                              co1: &CollisionObject<P, M, T>,
                              co2: &CollisionObject<P, M, T>);
}

A signal handler for contact detection.

Required Methods

fn handle_contact_started(&mut self,
co1: &CollisionObject<P, M, T>,
co2: &CollisionObject<P, M, T>,
contacts: &Box<ContactGenerator<P, M> + 'static>)

Activate an action for when two objects start being in contact.

fn handle_contact_stopped(&mut self,
co1: &CollisionObject<P, M, T>,
co2: &CollisionObject<P, M, T>)

Activate an action for when two objects stop being in contact.