Trait ncollide::narrow_phase::NarrowPhase

pub trait NarrowPhase<P, M, T> where P: Point {
    fn update(&mut self,
              objects: &UidRemap<CollisionObject<P, M, T>>,
              contact_signal: &mut ContactSignal<P, M, T>,
              proximity_signal: &mut ProximitySignal<P, M, T>,
              timestamp: usize);
    fn handle_interaction(&mut self,
                          contact_signal: &mut ContactSignal<P, M, T>,
                          proximity_signal: &mut ProximitySignal<P, M, T>,
                          objects: &UidRemap<CollisionObject<P, M, T>>,
                          fk1: &FastKey,
                          fk2: &FastKey,
                          started: bool);
    fn contact_pairs(&'a self,
                     objects: &'a UidRemap<CollisionObject<P, M, T>>)
                     -> ContactPairs<'a, P, M, T>;
    fn proximity_pairs(&'a self,
                       objects: &'a UidRemap<CollisionObject<P, M, T>>)
                       -> ProximityPairs<'a, P, M, T>;
}

Trait implemented by the narrow phase manager.

The narrow phase manager is responsible for creating, updating and generating contact pairs between objects identified by the broad phase.

Required Methods

fn update(&mut self,
objects: &UidRemap<CollisionObject<P, M, T>>,
contact_signal: &mut ContactSignal<P, M, T>,
proximity_signal: &mut ProximitySignal<P, M, T>,
timestamp: usize)

Updates this narrow phase.

fn handle_interaction(&mut self,
contact_signal: &mut ContactSignal<P, M, T>,
proximity_signal: &mut ProximitySignal<P, M, T>,
objects: &UidRemap<CollisionObject<P, M, T>>,
fk1: &FastKey,
fk2: &FastKey,
started: bool)

Called when the broad phase detects that two objects are, or stop to be, in close proximity.

fn contact_pairs(&'a self,
objects: &'a UidRemap<CollisionObject<P, M, T>>)
-> ContactPairs<'a, P, M, T>

Returns all the potential contact pairs found during the broad phase, and validated by the narrow phase.

fn proximity_pairs(&'a self,
objects: &'a UidRemap<CollisionObject<P, M, T>>)
-> ProximityPairs<'a, P, M, T>

Returns all the potential proximity pairs found during the broad phase, and validated by the narrow phase.