Struct ncollide::partitioning::BVT

pub struct BVT<B, BV> {
    // some fields omitted
}

A Boundig Volume Tree.

Methods

impl<B, BV> BVT<B, BV>

fn new_with_partitioner<F>(leaves: Vec<(B, BV)>, partitioner: &mut F) -> BVT<B, BV> where F: FnMut(usize, Vec<(B, BV)>) -> (BV, BinaryPartition<B, BV>)

Builds a bounding volume tree using an user-defined construction function.

fn visit<Vis>(&self, visitor: &mut Vis) where Vis: BVTVisitor<B, BV>

Traverses this tree using an object implementing the BVTVisitortrait.

This will traverse the whole tree and call the visitor .visit_internal(...) (resp. .visit_leaf(...)) method on each internal (resp. leaf) node.

fn visit_bvtt<Vis>(&self, other: &BVT<B, BV>, visitor: &mut Vis) where Vis: BVTTVisitor<B, BV>

Visits the bounding volume traversal tree implicitely formed with other.

fn best_first_search<N, BFS>(&'a self, algorithm: &mut BFS) -> Option<(&'a B, BFS::UserData)> where N: Scalar, BFS: BVTCostFn<N, B, BV>

Performs a best-fist-search on the tree.

Returns the content of the leaf with the smallest associated cost, and a result of user-defined type.

fn root_bounding_volume(&'r self) -> Option<&'r BV>

Reference to the bounding volume of the tree root.

fn depth(&self) -> usize

Computes the depth of this tree.

impl<B, BV> BVT<B, BV>

fn new_balanced<V>(leaves: Vec<(B, BV)>) -> BVT<B, BV> where BV: Translation<V> + BoundingVolume<V::Scalar> + Clone, V: Vector

Creates a balanced BVT.