Trait ncollide::partitioning::BVTCostFn

pub trait BVTCostFn<N, B, BV, R> {
    fn compute_bv_cost(&mut self, &BV) -> Option<N>;
    fn compute_b_cost(&mut self, &B) -> Option<(N, R)>;
}

Trait implemented by cost functions used by the best-first search on a BVT.

Required Methods

fn compute_bv_cost(&mut self, &BV) -> Option<N>

Computes the cost of a bounding volume.

fn compute_b_cost(&mut self, &B) -> Option<(N, R)>

Computes the cost of an object, and the result to be returned if it is the best one.