Module ncollide::point

Point inclusion and projection.

Structs

PointInterferencesCollector
Bounding Volume Tree visitor collecting nodes that may contain a given point.

Traits

LocalPointQuery
Trait of objects that can be tested for point inclusion and projection.

PointQuery
Trait of objects that can be transformed and tested for point inclusion and projection.