Function ncollide::procedural::unit_quad

pub fn unit_quad<P>(usubdivs: usize, vsubdivs: usize) -> TriMesh<P> where P: Point

Adds a double-sided quad with unit size to the scene.

The quad is initially centered at (0, 0, 0). Its normal is the z axis. The quad itself is composed of a user-defined number of triangles regularly spaced on a grid. This is the main way to draw height maps.

Arguments

usubdivs - number of horizontal subdivisions. This correspond to the number of squares which will be placed horizontally on each line. Must not be 0.

vsubdivs - number of vertical subdivisions. This correspond to the number of squares which will be placed vertically on each line. Must not be 0.