Enum ncollide::procedural::path::PathSample

pub enum PathSample<P> where P: Point {
    StartPoint(P, P::Vector),
    InnerPoint(P, P::Vector),
    EndPoint(P, P::Vector),
    EndOfSample,
}

A sample point and its associated tangent.

Variants
StartPoint(P, P::Vector)

A point that starts a new path.

InnerPoint(P, P::Vector)

A point that is inside of the path currently generated.

EndPoint(P, P::Vector)

A point that ends the path currently generated.

EndOfSample

Used when the sampler does not have any other points to generate.