Struct ncollide::procedural::path::PolylinePath

pub struct PolylinePath<'a, N, P, V> where P: 'a, V: 'a {
    // some fields omitted
}

A path with its sample points given by a polyline.

This will return sequencially each vertex of the polyline.

Methods

impl<'a, N, P, V> PolylinePath<'a, N, P, V> where N: Scalar, P: Point<N, V>, V: Vect<N>

fn new(polyline: &'a Polyline<N, P, V>) -> PolylinePath<'a, N, P, V>

Creates a new polyline-based path.