Trait ncollide::procedural::path::PolylineCompatibleCap

pub trait PolylineCompatibleCap<N> {
    fn gen_start_cap(&self, attach_id: u32, pattern: &Polyline<N, Pnt3<N>, Vec3<N>>, pt: &Pnt3<N>, dir: &Vec3<N>, closed: bool, coords: &mut Vec<Pnt3<N>>, indices: &mut Vec<Pnt3<u32>>);
    fn gen_end_cap(&self, attach_id: u32, pattern: &Polyline<N, Pnt3<N>, Vec3<N>>, pt: &Pnt3<N>, dir: &Vec3<N>, closed: bool, coords: &mut Vec<Pnt3<N>>, indices: &mut Vec<Pnt3<u32>>);
}

Trait to be implemented by caps compatible with a PolylinePattern.

Required Methods

fn gen_start_cap(&self, attach_id: u32, pattern: &Polyline<N, Pnt3<N>, Vec3<N>>, pt: &Pnt3<N>, dir: &Vec3<N>, closed: bool, coords: &mut Vec<Pnt3<N>>, indices: &mut Vec<Pnt3<u32>>)

Generates the mesh for the cap at the beginning of a path.

fn gen_end_cap(&self, attach_id: u32, pattern: &Polyline<N, Pnt3<N>, Vec3<N>>, pt: &Pnt3<N>, dir: &Vec3<N>, closed: bool, coords: &mut Vec<Pnt3<N>>, indices: &mut Vec<Pnt3<u32>>)

Generates the mesh for the cap at the end of a path.