Enum ncollide::query::algorithms::gjk::GJKResult

pub enum GJKResult<P, V> {
    Intersection,
    Projection(P),
    Proximity(V),
    NoIntersection(V),
}

Results of the GJK algorithm.

Variants
Intersection

Result of the GJK algorithm when the origin is inside of the polytope.

Projection(P)

Result of the GJK algorithm when a projection of the origin on the polytope is found.

Proximity(V)

Result of the GJK algorithm when the origin is to close to the polytope but not inside of it.

NoIntersection(V)

Result of the GJK algorithm when the origin is too far away from the polytope.