Function ncollide::query::ray_internal::triangle_ray_intersection

pub fn triangle_ray_intersection<P>(a: &P,
                                    b: &P,
                                    c: &P,
                                    ray: &Ray<P>)
                                    -> Option<(RayIntersection<P::Vector>, Matrix<P::Real, U3, U1, MatrixArray<P::Real, U3, U1>>)> where P: Point

Computes the intersection between a triangle and a ray.

If an intersection is found, the time of impact, the normal and the barycentric coordinates of the intersection point are returned.