Struct ncollide::query::ray_internal::Ray

pub struct Ray<P> where P: Point {
    pub origin: P,
    pub dir: P::Vector,
}

A Ray.

Fields
origin: P

Starting point of the ray.

dir: P::Vector

Direction of the ray.

Methods

impl<P> Ray<P> where P: Point

fn new(origin: P, dir: P::Vector) -> Ray<P>

Creates a new ray starting from origin and with the direction dir. dir must be normalized.

fn transform_by<M>(&self, m: &M) -> Ray<P> where M: Isometry<P>

Transforms this ray by the given isometry.

fn inverse_transform_by<M>(&self, m: &M) -> Ray<P> where M: Isometry<P>

Transforms this ray by the inverse of the given isometry.

fn translate_by(&self, v: P::Vector) -> Ray<P>

Translates this ray by the given vector. Its direction is left unchanged.