Struct ncollide::query::ray_internal::RayIntersectionCostFn

pub struct RayIntersectionCostFn<'a, P> where P: 'a + Point { /* fields omitted */ }

A search thet selects the objects that has the smallest time of impact with a given ray.

Methods

impl<'a, P> RayIntersectionCostFn<'a, P> where P: Point

fn new(ray: &'a Ray<P>, solid: bool, uvs: bool) -> RayIntersectionCostFn<'a, P>

Creates a new BestRayInterferenceSearch.