Trait ncollide::ray::RayCast

pub trait RayCast<P, M> where P: Point {
    fn toi_and_normal_with_ray(&self, m: &M, ray: &Ray<P>, solid: bool) -> Option<RayIntersection<P::Vect>>;

    fn toi_with_ray(&self, m: &M, ray: &Ray<P>, solid: bool) -> Option<P::Vect::Scalar> { ... }
    fn toi_and_normal_and_uv_with_ray(&self, m: &M, ray: &Ray<P>, solid: bool) -> Option<RayIntersection<P::Vect>> { ... }
    fn intersects_ray(&self, m: &M, ray: &Ray<P>) -> bool { ... }
}

Traits of objects which can be transformed and tested for intersection with a ray.

Required Methods

fn toi_and_normal_with_ray(&self, m: &M, ray: &Ray<P>, solid: bool) -> Option<RayIntersection<P::Vect>>

Computes the time of impact, and normal between this transformed shape and a ray.

Provided Methods

fn toi_with_ray(&self, m: &M, ray: &Ray<P>, solid: bool) -> Option<P::Vect::Scalar>

Computes the time of impact between this transform shape and a ray.

fn toi_and_normal_and_uv_with_ray(&self, m: &M, ray: &Ray<P>, solid: bool) -> Option<RayIntersection<P::Vect>>

Computes time of impact, normal, and texture coordinates (uv) between this transformed shape and a ray.

fn intersects_ray(&self, m: &M, ray: &Ray<P>) -> bool

Tests whether a ray intersects this transformed shape.