Module ncollide::shape

Collision shapes supported by ncollide.

Structs

AnnotatedMinkowskiSum
Same as the MinkowskiSum but with a support mapping which keeps track of the original supports points from the two wrapped shapes.

Ball
A Ball shape.

BaseMesh
A mesh generic wrt. the contained mesh elements characterized by vertices.

Capsule
SupportMap description of a capsule shape with its principal axis aligned with the y axis.

Compound
A compound shape with an aabb bounding volume.

Cone
SupportMap description of a cylinder shape with its principal axis aligned with the y axis.

Convex
The implicit convex hull of a set of points.

Cuboid
Shape of a box.

Cylinder
SupportMap description of a cylinder shape with its principal axis aligned with the y axis.

MinkowskiSum
SupportMap representation of the Minkowski sum of two shapes.

Plane
SupportMap description of a plane.

Polyline
Shape commonly known as a 2d line strip or a 3d segment mesh.

Reflection
SupportMap representation of the reflection of a shape.

Segment
A segment shape.

Torus
A torus.

TriMesh
Shape commonly known as a 2d line strip or a 3d triangle mesh.

Triangle
A triangle shape.

Traits

BaseMeshElement
Trait implemented by elements usable on the Mesh.

CompositeShape
Trait implemented by shapes composed of multiple simpler shapes.

Type Definitions

Ball2

Ball3

Capsule2

Capsule3

Compound2

Compound3

Cone2

Cone3

Convex2

Convex3

Cuboid2

Cuboid3

Cylinder2

Cylinder3

Plane2

Plane3

Polyline2

Polyline3

Segment2

Segment3

TriMesh3

Triangle2

Triangle3