Struct ncollide::shape::Compound

pub struct Compound<P, M> where P: Point { /* fields omitted */ }

A compound shape with an aabb bounding volume.

A compound shape is a shape composed of the union of several simpler shape. This is the main way of creating a concave shape from convex parts. Each parts can have its own delta transformation to shift or rotate it with regard to the other shapes.

Methods

impl<P, M> Compound<P, M> where M: Isometry<P>, P: Point

fn new(shapes: Vec<(M, ShapeHandle<P, M>)>) -> Compound<P, M>

Builds a new compound shape.

impl<P, M> Compound<P, M> where P: Point

fn shapes(&self) -> &[(M, ShapeHandle<P, M>)]

The shapes of this compound shape.

fn bvt(&self) -> &BVT<usize, AABB<P>>

The optimization structure used by this compound shape.

fn bounding_volumes(&self) -> &[AABB<P>]

The shapes bounding volumes.

fn aabb_at(&self, i: usize) -> &AABB<P>

The AABB of the i-th shape compositing this compound.