Struct ncollide::shape::Cone

pub struct Cone<N> {
    // some fields omitted
}

SupportMap description of a cylinder shape with its principal axis aligned with the y axis.

Methods

impl<N> Cone<N> where N: Scalar

fn new(half_height: N, radius: N) -> Cone<N>

Creates a new cone.

Arguments:

half_height - the half length of the cone along the y axis.

radius - the length of the cone along all other axis.

fn half_height(&self) -> N

The cone half length along the y axis.

fn radius(&self) -> N

The radius of the cone along all but the y axis.