Trait ncollide::shape::CompositeShape

pub trait CompositeShape<P, M> where P: Point {
    fn map_part_at(&self, usize, &mut FnMut(&M, &Repr<P, M> + 'static));
    fn map_transformed_part_at(&self, m: &M, usize, &mut FnMut(&M, &Repr<P, M> + 'static));
    fn aabb_at(&self, i: usize) -> &AABB<P>;
    fn bvt(&self) -> &BVT<usize, AABB<P>>;
}

Trait implemented by shapes composed of multiple simpler shapes.

A composite shape is composed of several shapes. Typically, it is a convex decomposition of a concave shape.

Required Methods

fn map_part_at(&self, usize, &mut FnMut(&M, &Repr<P, M> + 'static))

Applies a function to each sub-shape of this concave shape.

fn map_transformed_part_at(&self, m: &M, usize, &mut FnMut(&M, &Repr<P, M> + 'static))

Applies a transformation matrix and a function to each sub-shape of this concave shape.

fn aabb_at(&self, i: usize) -> &AABB<P>

Gets the AABB of the shape identified by the index i.

fn bvt(&self) -> &BVT<usize, AABB<P>>

Gets the acceleration structure of the concave shape.