Type Definition ncollide::shape::AnnotatedCSO

type AnnotatedCSO<'a, M, G1, G2> = AnnotatedMinkowskiSum<'a, M, G1, Reflection<'a, G2>>;

Type of an implicit representation of the Configuration Space Obstacle formed by two geometric objects. Uses return AnnotatedPoint instead of regular points, i.e., keeps tracks of the original points that yielded a given CSO support point.