Trait ncollide::support_map::PreferedSamplingDirections

pub trait PreferedSamplingDirections<V, M> {
    fn sample(&self, &M, &mut FnMut(V));
}

Trait of shapes having prefered sampling directions for the Minkowski sampling algorithm.

Those directions are usually the shape faces normals.

Required Methods

fn sample(&self, &M, &mut FnMut(V))

Applies a function to this shape with a given transform.