Module ncollide::transformation

Traits

ToPolyline
Trait implemented by shapes that can be approximated by a triangle mesh.

ToTriMesh
Trait implemented by shapes that can be approximated by a triangle mesh.

Functions

convex_hull2
Computes the convex hull of a set of 2d points.

convex_hull2_idx
Computes the convex hull of a set of 2d points and returns only the indices of the hull vertices.

convex_hull3
Computes the convex hull of a set of 3d points.

hacd
Approximate convex decomposition of a triangle mesh.

triangulate
Triangulates a set of point (sort of) lying on the same 2d plane.