Struct ncollide::utils::data::vec_slice::VecSlice

pub struct VecSlice<'a, T> where T: 'a {
    // some fields omitted
}

A vector slice with a specific length and stride.

The stride is the number of index increments between two elements.

Methods

impl<'a, T> VecSlice<'a, T>

fn new(data: &'a [T], length: usize, stride: usize) -> VecSlice<'a, T>

Creates a new immutable slice.

unsafe fn new_unsafe(data: &'a [T], length: usize, stride: usize) -> VecSlice<'a, T>

Creates a new immutable slice. The size of the data buffer is not checked.

fn len(&self) -> usize

The length of this slice.

fn is_empty(&self) -> bool

Whether or not this slice is empty.

fn get(&'b self, i: usize) -> &'b T

Gets the i-th element of the slice.

This is the same as the i * self.stride-th element of the wrapped vector.

unsafe fn get_unchecked(&'b self, i: usize) -> &'b T

Gets the i-th element of the slice without bound-checking.

This is the same as the i * self.stride-th element of the wrapped vector.