Function ncollide::utils::center_reduce

pub fn center_reduce<P>(pts: &mut [P]) -> (P::Vect::OuterProductType, P, bool) where P: Point, P::Vect: Vect, P::Vect: Outer, P::Vect::OuterProductType: Zero, P::Vect::OuterProductType: Mul<P>, P::Vect::OuterProductType: Inv, P::Vect::OuterProductType: Copy, P::Vect::OuterProductType: Debug, P::Vect::OuterProductType::Output == P

Centers and reduces a set of data.

Returns the covariance matrix, the center of the data, and a boolean that is true if the operation succeeded (otherwise, the returned value as valid, by the input points are left unchanged).