Function ncollide::utils::cross3

pub fn cross3<N, V>(a: &V, b: &V) -> V where V: Zero + Index<usize, Output=N, Output=N> + IndexMut<usize> + Dimension, N: Scalar

A 3d cross product that do not require the Cross<Self, Self> trait impl.

XXX: this is a HACK until the trait reform is done on rustc. This allows the use of the cross product without the need of the Cross<Self, Self> trait impl.

Fails if the dimension of V is not 3.