Struct ncollide::utils::HashablePartialEq

pub struct HashablePartialEq<T> {
    // some fields omitted
}

A structure that implements Eq and is hashable even if the wrapped data implements only PartialEq.

Methods

impl<T> HashablePartialEq<T>

unsafe fn new(value: T) -> HashablePartialEq<T>

Creates a new HashablePartialEq. This is unsafe because you must be sure that you really want to transform the wrapped object's partial equality to an equivalence relation.

fn unwrap(self) -> T

Gets the wrapped value.