Module ncollide::world

High level API to detect collisions in large, complex scenes.

Structs

CollisionGroups
Groups of collision used to filter which object interact with which other one.

CollisionGroupsPairFilter
A collision filter based collision groups.

CollisionObject
A stand-alone object that has a position and a shape.

CollisionWorld
A world that handles collision objects.

Enums

GeometricQueryType
The kind of query a CollisionObject may be involved on.

Type Definitions

BroadPhaseObject
Type of the broad phase trait-object used by the collision world.

CollisionObject2
A 2D collision object with user-defined data of type T.

CollisionObject3
A 3D collision object with user-defined data of type T.

CollisionWorld2
A 2D collision world associating collision objects to user-defined data of type T.

CollisionWorld3
A 3D collision world associating collision objects to user-defined data of type T.

NarrowPhaseObject
Type of the narrow phase trait-object used by the collision world.