Enum ncollide::world::GeometricQueryType

pub enum GeometricQueryType<N> where N: Real {
    Contacts(N),
    Proximity(N),
}

The kind of query a CollisionObject may be involved on.

The following queries are executed for a given pair of GeometricQueryType associated with two collision objects:

Contacts + Contacts = exact contact point coputation.

Contacts + Proximity = proximity test only.

Proximity + Proximity = proximity test only.

Variants
Contacts(N)

This objects can respond to both contact point computation and proximity queries.

Proximity(N)

This object can respond to proximity tests only.

Methods

impl<N> GeometricQueryType<N> where N: Real

fn query_limit(&self) -> N

The numerical limit of relevance for this query.

If two objects are separated by a distance greater than the sum of their respective query_limit, the corresponding query will not by performed. For proximity queries, non-intersecting object closer than a distance equal to the sum of their query_limit will be reported as Proximity::WithinMargin.

fn is_contacts_query(&self) -> bool

Returns true if this is a contacts query type.

fn is_proximity_query(&self) -> bool

Returns true if this is a proximity query type.