Struct ncollide::world::CollisionGroups

pub struct CollisionGroups {
    // some fields omitted
}

Groups of collision used to filter which object collide with which other one.

There are at most 30 groups indexed from 0 to 29 (included). This identifies collidable entities by combining three attributes: * A set of group this structure is member of. * A collision group whitelist. * A collision group blacklist.

For two entities to collide, they must be member of at least one group part of each-other's whitelists, and must not be part of any blacklisted group. The blacklist always has priority on the whitelist.

Example

For example if the object A is such that: * It is part of the groups 1, 3, and 6. * It whitelists the groups 3 and 7. * It blacklists the group 1. Let the object B be such that: * It is part of the groups 1, 3, and 7. * It whitelists the groups 3 and 7. * It does not blacklist anything. For example if the object C is such that: * It is part of the groups 6, 9. * It whitelists the groups 3 and 7. * It does not blacklist anything.

Then we have: * A and C can collide because A whitelists the group 6 (which C is part of), and, reciprocally, C whitelists the group 3 (which A is part of). * A and B will not collide because B is part of the group 1 which is blacklisted by A. * Finally, B and C will not collide either because, even if C whitelists the group 3 (which B is part of), B does not whitelists the groups 6 nor 9 (which B is part of).