Struct ncollide::world::CollisionObject

pub struct CollisionObject<P, M, T> where P: Point {
    pub uid: usize,
    pub position: M,
    pub shape: ShapeHandle<P, M>,
    pub collision_groups: CollisionGroups,
    pub query_type: GeometricQueryType<P::Vect::Scalar>,
    pub data: T,
    // some fields omitted
}

A stand-alone object that has a position and a shape.

Fields

uid

The collsion object unique identifier.

position

The collision object position.

shape

The collision object shape.

collision_groups

The collision groups of the collision object.

query_type

The kind of queries this collision object is expected to .

data

The user-defined data associated to this object.

Methods

impl<P, M, T> CollisionObject<P, M, T> where P: Point

fn new(uid: usize, position: M, shape: ShapeHandle<P, M>, groups: CollisionGroups, query_type: GeometricQueryType<P::Vect::Scalar>, data: T) -> CollisionObject<P, M, T>

Creates a new collision object.