Crate ncollide

ncollide

ncollide is a 2 and 3-dimensional collision detection library written with the rust programming language.

As its name suggests, it is generic wrt the dimension: it works with both 2-dimensional and 3-dimensional shapes. It might work with higher dimensions (never tried).

The official user guide is available here. The rustdoc documentation is available here.

Compilation

You will need the last nightly build of the rust compiler and the official package manager: cargo.

Simply add the following to your Cargo.toml file:

[dependencies.ncollide]
git = "https://github.com/sebcrozet/ncollide"

Features

dynamic bounding volume tree based broad phase

ball vs. ball collision detection,

plane vs. any convex object collision detection.

collision detection between arbitrary convex objects

compound shapes

ray-casting

time of impact computation for objects without rotational movement (compound vs. compound is not yet implemented)

And various traits for collision detectors and broad phase collision detection.

Modules

bounding_volume
Bounding volumes.

broad_phase
Broad phases.

geometry
Non-persistant pairwise geometric queries.

inspection
Traits and methods to inspect and retrieve the capabilities and representations of shapes at runtime.

math

narrow_phase
Collision detection algorithms and structure for the Narrow Phase.

partitioning
Spatial partitioning tools.

point
Point inclusion and projection.

procedural

ray
Ray-casting related definitions and implementations.

shape
Collision shapes supported by ncollide.

support_map
Definition of support functions.

transformation

utils

world
High level API to detect collisions in large, complex scenes.