Struct ncollide::bounding_volume::AABB

pub struct AABB<P> {
    // some fields omitted
}

An Axis Aligned Bounding Box.

Methods

impl<P> AABB<P> where P: POrd

fn new(mins: P, maxs: P) -> AABB<P>

Creates a new AABB.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

impl<P> AABB<P> where P: Neg, P: Bounded, P: POrd, <P as Neg>::Output == P

fn new_invalid() -> AABB<P>

Creates an invalid AABB with: * mins = Bounded::max_value() * maxs = Bounded::max_value(). This is useful to build aabb using merges.

impl<P> AABB<P>

fn mins(&self) -> &P

Reference to the AABB point with the smallest components along each axis.

fn maxs(&self) -> &P

Reference to the AABB point with the biggest components along each axis.

impl<N, P, V> AABB<P> where N: Scalar, P: Point<N, V>, V: Vect<N>

fn center(&self) -> P

The center of this AABB.

fn half_extents(&self) -> V

The half extents of this AABB.