Struct ncollide::bounding_volume::AABB

pub struct AABB<P> { /* fields omitted */ }

An Axis Aligned Bounding Box.

Methods

impl<P> AABB<P> where P: Point

fn new(mins: P, maxs: P) -> AABB<P>

Creates a new AABB.

Arguments:

mins - position of the point with the smallest coordinates.

maxs - position of the point with the highest coordinates. Each component of mins must be smaller than the related components of maxs.

fn mins(&self) -> &P

Reference to the AABB point with the smallest components along each axis.

fn maxs(&self) -> &P

Reference to the AABB point with the biggest components along each axis.

fn center(&self) -> P

The center of this AABB.

fn half_extents(&self) -> P::Vector

The half extents of this AABB.