Module ncollide::geometry::distance_internal

Implementation details of the distance function.

Functions

any_against_any
Computes the minimum distance separating two shapes.

any_against_any
Computes the minimum distance separating two shapes.

any_against_composite_shape
Smallest distance between a shape and a composite shape.

ball_against_ball
Distance between balls.

composite_shape_against_any
Smallest distance between a composite shape and any other shape.

plane_against_support_map
Distance between a plane and a support-mapped shape.

support_map_against_plane
Distance between a support-mapped shape and a plane.

support_map_against_support_map
Distance between support-mapped shapes.

support_map_against_support_map_with_params
Distance between support-mapped shapes.