Module ncollide::narrow_phase

Persistant collision detection algorithms to compute contact points.

Structs

BallBallCollisionDetector
Collision detector between two balls.

BallBallProximityDetector
Proximity detector between two balls.

CompositeShapeReprCollisionDetector
Collision detector between a concave shape and another shape.

CompositeShapeReprProximityDetector
Proximity detector between a concave shape and another shape.

ContactPairs
Iterator through contact pairs.

ContactSignal
Signal for contact start/stop.

Contacts
An iterator through contacts.

DefaultCollisionDispatcher
Collision dispatcher for shapes defined by ncollide_entities.

DefaultNarrowPhase
Collision detector dispatcher for collision objects.

DefaultProximityDispatcher
Proximity dispatcher for shapes defined by ncollide_entities.

IncrementalContactManifoldGenerator
Contact manifold generator which keeps track of several contacts.

OneShotContactManifoldGenerator
Contact manifold generator producing a full manifold at the first update.

PlaneSupportMapCollisionDetector
Collision detector between a plane and a shape implementing the SupportMap trait.

PlaneSupportMapProximityDetector
Proximity detector between a plane and a shape implementing the SupportMap trait.

ProximitySignal
Signal for proximity start/stop.

ReprCompositeShapeCollisionDetector
Collision detector between a shape and a concave shape.

ReprCompositeShapeProximityDetector
Proximity detector between a shape and a concave shape.

SupportMapPlaneCollisionDetector
Collision detector between a plane and a shape implementing the SupportMap trait.

SupportMapPlaneProximityDetector
Proximity detector between a plane and a shape implementing the SupportMap trait.

SupportMapSupportMapCollisionDetector
Persistent collision detector between two shapes having a support mapping function.

SupportMapSupportMapProximityDetector
Persistent proximity detector between two shapes having a support mapping function.

Traits

CollisionDetector
Trait implemented algorithms that compute contact points, normals and penetration depths.

CollisionDispatcher

ContactSignalHandler
A signal handler for contact detection.

NarrowPhase
Trait implemented by the narrow phase manager.

ProximityDetector
Trait implemented by algorithms that determine if two objects are in close proximity.

ProximityDispatcher

ProximitySignalHandler
A signal handler for proximity detection.

Type Definitions

CollisionAlgorithm

ProximityAlgorithm