Trait ncollide::point::PointQuery

pub trait PointQuery<P, M> where P: Point {
    fn project_point(&self, m: &M, pt: &P, solid: bool) -> P;
    fn distance_to_point(&self, m: &M, pt: &P) -> P::Vect::Scalar;
    fn contains_point(&self, m: &M, pt: &P) -> bool;
}

Trait of objects that can be tested for point inclusion and projection.

Required Methods

fn project_point(&self, m: &M, pt: &P, solid: bool) -> P

Projects a point on self transformed by m.

fn distance_to_point(&self, m: &M, pt: &P) -> P::Vect::Scalar

Computes the minimal distance between a point and self transformed by m.

fn contains_point(&self, m: &M, pt: &P) -> bool

Tests if the given point is inside of self transformed by m.