Struct ncollide::procedural::path::PolylinePath

pub struct PolylinePath<'a, P> where P: 'a + Point {
    // some fields omitted
}

A path with its sample points given by a polyline.

This will return sequencially each vertex of the polyline.

Methods

impl<'a, P> PolylinePath<'a, P> where P: Point

fn new(polyline: &'a Polyline<P>) -> PolylinePath<'a, P>

Creates a new polyline-based path.