Function ncollide::ray::triangle_ray_intersection

pub fn triangle_ray_intersection<P>(a: &P, b: &P, c: &P, ray: &Ray<P>) -> Option<(RayIntersection<P::Vect>, Vec3<P::Vect::Scalar>)> where P: Point

Computes the intersection between a triangle and a ray.

If an intersection is found, the time of impact, the normal and the barycentric coordinates of the intersection point are returned.