Struct ncollide::ray::RayIntersection

pub struct RayIntersection<N, V> {
    pub toi: N,
    pub normal: V,
    pub uvs: Option<Pnt2<N>>,
}

Structure containing the result of a successful ray cast.

Fields

toi

The time of impact of the ray with the object. The exact contact point can be computed with: orig + dir * toi where orig is the origin of the ray; dir is its direction and toi is the value of this field.

normal

The normal at the intersection point.

If the toi is exactly zero, the normal might not be reliable.

uvs

The textures coordinates at the intersection point. This is an Option because some shape do not support texture coordinates.

Methods

impl<N, V> RayIntersection<N, V>

fn new_with_uvs(toi: N, normal: V, uvs: Option<Pnt2<N>>) -> RayIntersection<N, V>

Creates a new RayIntersection.

fn new(toi: N, normal: V) -> RayIntersection<N, V>

Creates a new RayIntersection.