Struct ncollide::ray::RayIntersectionCostFn

pub struct RayIntersectionCostFn<'a, N, P, V> where P: 'a, V: 'a {
    // some fields omitted
}

A search thet selects the objects that has the smallest time of impact with a given ray.

Methods

impl<'a, N, P, V> RayIntersectionCostFn<'a, N, P, V>

fn new(ray: &'a Ray<P, V>, solid: bool, uvs: bool) -> RayIntersectionCostFn<'a, N, P, V>

Creates a new BestRayInterferenceSearch.