Trait ncollide::ray::RayCast

pub trait RayCast<N, P, V, M>: LocalRayCast<N, P, V> where M: Rotate<V>, M: Transform<P> {
    fn toi_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>, solid: bool) -> Option<N> { ... }
    fn toi_and_normal_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>> { ... }
    fn toi_and_normal_and_uv_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>> { ... }
    fn intersects_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>) -> bool { ... }
}

Traits of objects which can be transformed and tested for intersection with a ray.

Provided Methods

fn toi_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>, solid: bool) -> Option<N>

Computes the time of impact between this transform shape and a ray.

fn toi_and_normal_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>>

Computes the time of impact, and normal between this transformed shape and a ray.

fn toi_and_normal_and_uv_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>, solid: bool) -> Option<RayIntersection<N, V>>

Computes time of impact, normal, and texture coordinates (uv) between this transformed shape and a ray.

fn intersects_with_transform_and_ray(&self, m: &M, ray: &Ray<P, V>) -> bool

Tests whether a ray intersects this transformed shape.