Struct ncollide::shape::AnnotatedMinkowskiSum

pub struct AnnotatedMinkowskiSum<'a, M, G1, G2> where G1: 'a + ?Sized, M: 'a, G2: 'a + ?Sized {
    // some fields omitted
}

Same as the MinkowskiSum but with a support mapping which keeps track of the original supports points from the two wrapped shapes.

G1: type of the first object involved on the sum.

G2: type of the second object involved on the sum.

Methods

impl<'a, M, G1, G2> AnnotatedMinkowskiSum<'a, M, G1, G2> where G1: ?Sized, G2: ?Sized

fn new(m1: &'a M, g1: &'a G1, m2: &'a M, g2: &'a G2) -> AnnotatedMinkowskiSum<'a, M, G1, G2>

Builds the Minkowski sum of two shapes. Since the representation is implicit, this is done in constant time.

fn m1(&self) -> &'a M

The transformation matrix of the first shape of this Minkowski Sum.

fn m2(&self) -> &'a M

The transformation matrix of the second shape of this Minkowski Sum.

fn g1(&self) -> &'a G1

The first shape of this Minkowski Sum.

fn g2(&self) -> &'a G2

The second shape of this Minkowski Sum.