Struct ncollide::shape::Compound

pub struct Compound<N, P, V, M> {
    // some fields omitted
}

A compound shape with an aabb bounding volume.

A compound shape is a shape composed of the union of several simpler shape. This is the main way of creating a concave shape from convex parts. Each parts can have its own delta transformation to shift or rotate it with regard to the other shapes.

Methods

impl<N, P, V, M> Compound<N, P, V, M> where N: Scalar, P: Point<N, V>, V: Vect<N>, V: Translate<P>, M: Isometry<N, P, V>

fn new(shapes: Vec<(M, Arc<Box<Repr<N, P, V, M> + 'static>>)>) -> Compound<N, P, V, M>

Builds a new compound shape.

impl<N, P, V, M> Compound<N, P, V, M>

fn shapes(&self) -> &[(M, Arc<Box<Repr<N, P, V, M> + 'static>>)]

The shapes of this compound shape.

fn bvt(&self) -> &BVT<usize, AABB<P>>

The optimization structure used by this compound shape.

fn bounding_volumes(&self) -> &[AABB<P>]

The shapes bounding volumes.

fn aabb_at(&self, i: usize) -> &AABB<P>

The AABB of the i-th shape compositing this compound.