Module ncollide::utils::data

Data structure utilities.

Modules

hash
Trait for hash functions.

hash_map
An hash map with a customizable hash function.

owned_allocation_cache
Allocation cache for owned objects.

pair
Hashable pair of objects implementing HasUid.

ref_with_cost
A reference packed with a cost value.

uid_remap
A map allowing a slow lookup for arbitrary usize and fast lookup for small ones.

vec_map
A simple map based on a vector for small integer keys. Space requirements are O(highest integer key).

vec_slice
Slicing on non-contiguous data.