Struct ncollide::world::CollisionGroups

pub struct CollisionGroups {
    // some fields omitted
}

30 groups of collision used to filter which object collide with which other one.

If two objects have at least one collision group in common, they can collide. Otherwise, they cannot.